Initial release of LoungeLooker
Using a Raspberry Pi 4 running OpenCV with a webcam for facial-recognition, the software (in Python) claims to calculate your innermost desirees, and then chooses the perfect lounge music song (played on three of my ArduTouch synthesizers, plus the open source eSpeak text-to-speech engine for lyrics) to fulfill those innermost desires.

I developed this project as Artist-in-Residence with Monochrome/Q21 at MuseumsQuartier in Vienna in August/September-2021, and presented at Roboexotica as part of Ars Electronica in September-2021.